+++ docs/CHANGELOG.rst
@@ -40,6 +40,7 @@ Fixed
 - Fix ``--cl-config`` input option in MultiQC process. Use
   ``resolwebio/common:3.0.1`` Docker image with updated MultiQC
   configuration file to omit parsing the unwanted ``tmp`` folder
+- Fix LISTENER_CONNECTION settings to work on Mac
 
 
 ===================